Move Layers into r3f-stage
