+++ packages/r3f-stage/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./Layers"
 export * from "./R3FStage"
